requirements.txt
PyMuPDF==1.24.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
transformers==4.43.2
scikit-learn==1.4.2
xgboost==2.0.3
numpy==1.26.4
tqdm==4.66.4
